+++ lib/djv/App/App.h
@@ -334,6 +334,7 @@ namespace djv
             void _reloadUpdate(const std::shared_ptr<models::FilesModelItem>&);
             void _layersUpdate(const std::vector<int>&);
             void _mediaReferenceKeysUpdate(const std::vector<std::string>&);
+            void _mediaReferenceWarning(const models::FilesModelItem&, const std::string&);
             void _audioUpdate();
 
             void _reviewFileDialog(